+++ CHANGELOG.md
@@ -7,6 +7,22 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Fixed
 - Use strict Base64 encoding to fix base64 encoding/netty issue (@msblum)
 
+## [1.2.0] - 2017-05-03
+### Fixed
+- metrics-es-cluster.rb: Check to see if cluster key exists in transient_settings before trying to use it. (@RAR)
+### Added
+- Add option to run check-es-shard-allocation-status.rb on non master nodes (@Evesy)
+- Fixed check-es-shard-allocation-status.rb for Elasticsearch 5.x compatibility (@Evesy)
+
+## [1.1.3] - 2017-01-04
+### Fixed
+- metrics-es-cluster/metrics-es-node-graphite.rb: Fix Elasticsearch 5.0 compatability (@terjesannum)
+
+## [1.1.2] - 2016-12-29
+### Fixed
+- Fixed metrics-es-node-graphite.rb was not compatible with Elasticsearch 5.0 (@woqer)
+- Make query lib compatible with ES 5+ (@jackfengji)
+
 ## [1.1.1] - 2016-11-26
 ### Fixed
 - Fixed check-es-file-descriptors.rb was not compatible with Elasticsearch 5.0 (@woqer)
@@ -128,7 +144,10 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Added
 - initial release
 
-[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/1.1.1...HEAD
+[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/1.1.3...1.2.0
+[1.1.3]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/1.1.2...1.1.3
+[1.1.2]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/sensu-plugins/sensu-plugins-elasticsearch/compare/0.5.3...1.0.0